requirements.txt
Django==3.2.4
djangorestframework==3.12.4
dadata==21.10.1